// src/app/core/model/product.model.ts
export interface Product {
    id: number
    title: string
    description: string
    price: number
    discountPercentage: number
    rating: number
    stock: number
    brand: string
    category: string
    thumbnail: string
    images: string[]
    isDeleted?: boolean
    deletedOn?: boolean
}